Implement JUnit 5 test extension in Paparazzi class
